Update ft_putnbr_base.c 1807

+++ 2try/ft_putnbr_base.c
@@ -35,7 +35,7 @@ size_t	print_base(ssize_t num, char *base, size_t base_len)
 	return (count);
 }
 
-size_t	ft_putnbr_base(ssize_t nbr, char *base, size_t base_len, char spf)
+int	ft_putnbr_base(ssize_t nbr, char *base, size_t base_len, char spf)
 {
 	size_t	count;
